Restore Stat and Expr to be aliases of UInt (not UInt8)
This reverts a change from 2014-10-29, which is no longer necessary
due to the introduction of struct StatHeader.

+++ src/code.h
@@ -29,7 +29,7 @@
 **  If 'Stat' is different  from 'Expr', then  a lot of things will  probably
 **  break.
 */
-typedef UInt8 Stat;
+typedef UInt Stat;
 
 typedef struct {
     unsigned int visited : 1;
